Deletion functionality complete. Integrated with backend.
Agent 1 is the only agent not able to be deleted due to MySQL/Internal server error (i.e. error code 500 is displayed in the console because you cannot delete a parent row (agent 1) before deleting a child row (their alias)). This was checked by performing deletion queries on agent 1 in MySQL workbench.